requirements.txt
# MSSQL to dbt Migration Tool - Requirements

# Core dependencies
anthropic>=0.39.0
python-dotenv>=1.0.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.40
pydantic>=2.0.0

# AWS
boto3>=1.34.0
aws-cdk-lib>=2.120.0
constructs>=10.0.0
moto>=5.0.0

# Database connectivity
pyodbc>=5.0.0
sqlalchemy>=2.0.0

# dbt
dbt-core>=1.7.0
dbt-duckdb>=1.7.0  # Using DuckDB for POC simplicity

# Utilities
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
click>=8.1.0
networkx>=3.0  # For dependency graph analysis

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
